Add URL logger to download and upload methods
This feature is opt-out when used from the CLI, and opt-in
when used from the library interface. It is not intended for
logging error messages. Each line adheres to the following formaating:

DATE::METHOD::URL

where :: serves as a delimiter between those fields. For example:

2021-08-21 00:26:42::upload::https://anonfiles.com/jfw1h7Dcu7/logo_svg

The get_config_dir defines the destination of the configuration files
which is platform-specific and best understood by reading its
implementation.