Fix StopIteration error after parsing a URL
When trying to download or preview a file that AnonFile detects as malware, anonfile-api
does not match because it search between URLS of an href attribute, but the URL is of a
value attribute.

AnonFile seems to scan a file when its size is probably <= 30 MiB. Sometimes it detects
some files as malware, but I check them in VirusTotal and there is no problem. I think
users who download such files should be their own responsible, so it should be
downloaded without user intervention.